refactor: move Result type to a dedicated file
Centralized the Result interface in src/types/results.ts to enhance modularity.
