+++ src/parser/processor.ts
@@ -1,10 +1,10 @@
 import Decimal from "decimal.js";
 import * as fs from "fs";
-import { CommentAssociation, CommentKind } from "../configuration/comment-types";
 import { typeReplacer } from "../helpers/result-replacer";
 import { IssueActivity } from "../issue-activity";
 import { Module } from "../types/module";
 import { ContextPlugin } from "../types/plugin-input";
+import { Result } from "../types/results";
 import { ContentEvaluatorModule } from "./content-evaluator-module";
 import { DataPurgeModule } from "./data-purge-module";
 import { FormattingEvaluatorModule } from "./formatting-evaluator-module";
@@ -71,44 +71,4 @@ export class Processor {
 
     return totalReward.toNumber();
   }
-}
-
-export interface Result {
-  [k: string]: {
-    comments?: GithubCommentScore[];
-    total: number;
-    task?: {
-      reward: number;
-      multiplier: number;
-    };
-    feeRate?: number;
-    permitUrl?: string;
-    userId: number;
-    evaluationCommentHtml?: string;
-  };
-}
-
-export interface WordResult {
-  wordCount: number;
-  wordValue: number;
-  result: number;
-}
-
-export interface GithubCommentScore {
-  id: number;
-  content: string;
-  url: string;
-  type: CommentKind | CommentAssociation;
-  diffHunk?: string;
-  score?: {
-    formatting?: {
-      content: Record<string, { score: number; elementCount: number }>;
-      result: number;
-    };
-    words?: WordResult;
-    multiplier: number;
-    relevance?: number;
-    clarity?: number;
-    reward: number;
-  };
 }
