+++ src/run.ts
@@ -3,9 +3,10 @@ import { collectLinkedMergedPulls } from "./data-collection/collect-linked-pulls
 import { GITHUB_DISPATCH_PAYLOAD_LIMIT } from "./helpers/constants";
 import { getSortedPrices } from "./helpers/label-price-extractor";
 import { IssueActivity } from "./issue-activity";
-import { Processor, Result } from "./parser/processor";
+import { Processor } from "./parser/processor";
 import { parseGitHubUrl } from "./start";
 import { ContextPlugin } from "./types/plugin-input";
+import { Result } from "./types/results";
 
 export async function run(context: ContextPlugin) {
   const { eventName, payload, logger, config } = context;
