+++ tests/parser/permit-generation-module.test.ts
@@ -4,9 +4,9 @@ import { paginateGraphQL } from "@octokit/plugin-paginate-graphql";
 import { Octokit } from "@octokit/rest";
 import { Logs } from "@ubiquity-os/ubiquity-os-logger";
 import { CommentKind } from "../../src/configuration/comment-types";
-import { Result } from "../../src/parser/processor";
 import { EnvConfig } from "../../src/types/env-type";
 import { ContextPlugin } from "../../src/types/plugin-input";
+import { Result } from "../../src/types/results";
 import { db } from "../__mocks__/db";
 import dbSeed from "../__mocks__/db-seed.json";
 import { server } from "../__mocks__/node";
